refactor(CLIPull): moving project download to utils
This is better for code reuse.
